backend/requirements.txt
fastapi
uvicorn
python-socketio
numpy
